+++ src/me/dylan/mvcGame/game/gameObjects/Tiles.java
@@ -1,9 +1,9 @@
 package me.dylan.mvcGame.game.gameObjects;
 
 public class Tiles {
-    public static final int SPECIAL_ID = 0;
-    public static final int FLOOR_ID = 1;
-    public static final int WALL_ID = 2;
-    public static final int START_ID = 3; //TODO make a special tile for the start
-    public static final int END_ID = 4; //TODO make a special tile for the end
+    public static final byte SPECIAL_ID = 0;
+    public static final byte FLOOR_ID = 1;
+    public static final byte WALL_ID = 2;
+    public static final byte START_ID = 3; //TODO make a special tile for the start
+    public static final byte END_ID = 4; //TODO make a special tile for the end
 }
